Create unitConversion utility file
Move the existing `convertMegabytesTo` function to a new utility file. Create a new function called `readableBytes`, which takes in a number in bytes and returns a human-readable version.

This change was originally written in the closed PR, #4857. It was cleaner to re-write it instead of cherry picking a partial commit.
